Remove swear word thing
